requirements.txt
kaggle==1.7.4.2
python-dotenv==1.1.0
numpy==2.2.4
pandas==2.2.3
matplotlib==3.10.1